DEP/MAINT: Deprecate old models, refactor into one model.
